requirements.txt
selenium==3.141.0
Appium-Python-Client
pytest
allure-pytest
jprops